movie_review_project/requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.8
importlib_resources==6.4.4
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==1.24.4
pandas==2.0.3
pkgutil_resolve_name==1.3.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-surprise==1.1.4
scipy==1.10.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
zipp==3.20.1
